requirements.txt
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
numpy>=1.26.0 
scikit-learn>=1.3.0
scipy>=1.10.0
pytest>=7.4.0
jupyter>=1.0.0
eventregistry